Move methods for reading notes to NoteList package; Fix bug when creating new notes
